chore: delete unused views
