--- Oscar°/ViewModels/SearchViewModel.swift
@@ -1,70 +1,0 @@
-//
-//  SearchViewModel.swift
-//  SearchViewModel
-//
-//  Created by Philipp Bolte on 16.08.21.
-//  Credits to https://www.peteralt.com/blog/mapkit-location-search-with-swiftui/
-//
-
-import Foundation
-import Combine
-import MapKit
-
-class SearchViewModel: NSObject, ObservableObject {
-
-    enum SearchStatus: Equatable {
-        case idle
-        case noResults
-        case isSearching
-        case error(String)
-        case result
-    }
-
-    @Published var queryFragment: String = ""
-    @Published var status: SearchStatus = .idle
-    @Published private(set) var searchResults: [MKLocalSearchCompletion] = []
-
-    private var queryCancellable: AnyCancellable?
-    private let searchCompleter: MKLocalSearchCompleter!
-
-    init(searchCompleter: MKLocalSearchCompleter = MKLocalSearchCompleter()) {
-        self.searchCompleter = searchCompleter
-        super.init()
-        self.searchCompleter.delegate = self
-
-        queryCancellable = $queryFragment
-            .receive(on: DispatchQueue.main)
-            .debounce(for: .milliseconds(150), scheduler: RunLoop.main, options: nil)
-            .sink(receiveValue: { fragment in
-                self.status = .isSearching
-                if !fragment.isEmpty {
-                    self.searchCompleter.queryFragment = fragment
-                } else {
-                    self.status = .idle
-                    self.searchResults = []
-                }
-        })
-    }
-    
-    func getCoordinates(searchCompletion: MKLocalSearchCompletion, completion: @escaping (CLLocationCoordinate2D) -> Void) {
-        let searchRequest = MKLocalSearch.Request(completion: searchCompletion)
-        let search = MKLocalSearch(request: searchRequest)
-        
-        search.start { response, error in
-            let coordinates = response?.mapItems[0].placemark.coordinate ?? CLLocationCoordinate2D(latitude: 0, longitude: 0)
-            completion(coordinates)
-        }
-    }
-}
-
-extension SearchViewModel: MKLocalSearchCompleterDelegate {
-    func completerDidUpdateResults(_ completer: MKLocalSearchCompleter) {
-        self.searchResults = completer.results
-        self.status = completer.results.isEmpty ? .noResults : .result
-    }
-
-    func completer(_ completer: MKLocalSearchCompleter, didFailWithError error: Error) {
-        self.status = .error(error.localizedDescription)
-    }
-}
-
