+++ Oscar°/Views/RainRadarView.swift
@@ -47,54 +47,3 @@ struct RainRadarView: UIViewRepresentable {
         
     }
 }
-
-struct ContentView2 : View {
-    
-    @State private var timer : Timer?
-    @State private var overlay : MKTileOverlay?
-    @State private var pos = 0
-    @State private var time: String?
-    @Binding var radarMetadata: WeatherMapsResponse?
-    
-    var body : some View {
-        
-        VStack{
-            RainRadarView(timer: $timer, overlay: $overlay)
-            Text(time ?? "")
-            
-            Button(action:{
-                if self.timer == nil{
-                    self.timer = Timer.scheduledTimer(withTimeInterval : 1 ,repeats:true){_ in
-                    
-                        let path = radarMetadata?.radar.past[pos].path
-                        self.time = Date(timeIntervalSince1970: TimeInterval(radarMetadata?.radar.past[pos].time ?? 0)).formatted()
-                        
-                        // Create rain radar tile overlay with date parameter
-                        let template = "https://tilecache.rainviewer.com\(path ?? "")/256/{z}/{x}/{y}/2/1_0.png"
-                        let overlay = MKTileOverlay(urlTemplate : template)
-                        
-                        // Remove previous overlay if any and add new one
-                        self.overlay = overlay
-
-                        if (pos == (radarMetadata?.radar.past.count ?? 0 )-1) {
-                            pos = 0
-                        } else {
-                            pos += 1
-                        }
-                    }
-                }else{
-                    self.timer?.invalidate()
-                    self.timer=nil
-                    
-                }
-            }){
-                Text(self.timer == nil ? "Play" : "Stop")
-                    .font(.largeTitle)
-                
-            }
-            
-        }
-        
-    }
-    
-}
